Add part for notes and improve gitstat formatting
Prepend an empty plain text part to the start of the mail body to make it easy
to include some notes about the patch prior to sending it out for review.

To distinguish the notes from the gitstat section, wrap the gitstat in pre
tags within an HTML part, ensuring that the body is HTML escaped.